* updated Input so that cursor control by keyboard will adjust its speed    according to current FPS. (header)
git-svn-id: https://shooting-cubes.googlecode.com/svn@296 93678846-683e-0410-93c7-9b11d4a5ef47

+++ trunk/include/Input.hpp
@@ -104,7 +104,7 @@ private:
     int haste_key_;
     int pause_key_;
 
-    int CURSOR_SENSATIVITY;
+    float CURSOR_SENSATIVITY;
 
     wpPlayer player_;
     view::pSprite cursor_mark_;
